Refactoring - gathering all common for client & server classes in the new module, which is dependency for client and server. + Reworking DiskWorker (local as main & remote as child) to avoid code doubling.
